+++ mypy/newsemanal/semanal_classprop.py
@@ -1,6 +1,6 @@
-"""Calculate the abstract status of classes.
+"""Calculate some properties of classes.
 
-This happens after semantic analysis and before type checking.
+These happen after semantic analysis and before type checking.
 """
 
 from typing import List, Set, Optional
@@ -61,3 +61,24 @@ def calculate_class_abstract_status(typ: TypeInfo, is_stub_file: bool, errors: E
             report("Class {} has abstract attributes {}".format(typ.fullname(), attrs), 'error')
             report("If it is meant to be abstract, add 'abc.ABCMeta' as an explicit metaclass",
                    'note')
+
+
+def calculate_class_vars(info: TypeInfo) -> None:
+    """Try to infer additional class variables.
+
+    Subclass attribute assignments with no type annotation are assumed
+    to be classvar if overriding a declared classvar from the base
+    class.
+
+    This must happen after the main semantic analysis pass, since
+    this depends on base class bodies having been fully analyzed.
+    """
+    for name, sym in info.names.items():
+        node = sym.node
+        if isinstance(node, Var) and node.info and node.is_inferred and not node.is_classvar:
+            for base in info.mro[1:]:
+                member = base.names.get(name)
+                if (member is not None
+                        and isinstance(member.node, Var)
+                        and member.node.is_classvar):
+                    node.is_classvar = True
